Fixed bug in settings dialog TVInmages and TVInfo
Due to panel managemnt the property tabs were not visible.
Updated and fixed order.
Also visibility flag is updated correct AND size of the window. Panels
now are visible when editing the dialog and not overlapped to avoid
mistakes. Space on the right of main panel is free to move and edit sub
panels.