Updated project structure with additional needed files
Updated all project structures to include the addtional files in the
correct folders so that the build process will put them in the right
place.
This is needed to have languages in the correct folders.
DanCooper: please import this and update the structure.
I've discovered there are items missing in the language files and will
fix them but I must have this structure in place